fastAPI/requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
idna==3.4
imageio==2.30.0
imageio-ffmpeg==0.4.8
minio==7.1.15
moviepy==1.0.3
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
